refactor: conform page
