+++ apps/example-next-conform/src/app/forms/[storeName]/static-form/schema.ts
