+++ apps/example-next-conform/src/app/page.tsx
@@ -11,22 +11,28 @@ export default function Home() {
       <h1>`conform` example</h1>
       <ul>
         <li>
-          <Link href="/session/simple-form">/session/simple-form</Link>
+          <Link href="/forms/session/simple-form">
+            /forms/session/simple-form
+          </Link>
         </li>
         <li>
-          <Link href="/session/static-form">/session/static-form</Link>
+          <Link href="/forms/session/static-form">
+            /forms/session/static-form
+          </Link>
         </li>
         <li>
-          <Link href="/session/dynamic-form">/session/dynamic-form</Link>
+          <Link href="/forms/session/dynamic-form">
+            /forms/session/dynamic-form
+          </Link>
         </li>
         <li>
-          <Link href="/url/simple-form">/url/simple-form</Link>
+          <Link href="/forms/url/simple-form">/forms/url/simple-form</Link>
         </li>
         <li>
-          <Link href="/url/static-form">/url/static-form</Link>
+          <Link href="/forms/url/static-form">/forms/url/static-form</Link>
         </li>
         <li>
-          <Link href="/url/dynamic-form">/url/dynamic-form</Link>
+          <Link href="/forms/url/dynamic-form">/forms/url/dynamic-form</Link>
         </li>
       </ul>
     </main>
