--- apps/example-next-conform/src/app/url/simple-form/page.tsx
@@ -1,10 +1,0 @@
-import Form from "./form";
-
-export default function Page() {
-  return (
-    <main>
-      <h1>Simple Form</h1>
-      <Form />
-    </main>
-  );
-}
