--- apps/example-next-conform/src/app/url/static-form/page.tsx
@@ -1,10 +1,0 @@
-import Form from "./form";
-
-export default function Page() {
-  return (
-    <main>
-      <h1>Static Form</h1>
-      <Form />
-    </main>
-  );
-}
